refactor: re-organize onestore options
