+++ internal/database/metadata/database.go
@@ -6,6 +6,7 @@ import (
 	"io"
 
 	"github.com/onestore-ai/onestore/internal/database/metadata/postgres"
+	dbtypes "github.com/onestore-ai/onestore/internal/database/types"
 	"github.com/onestore-ai/onestore/pkg/onestore/types"
 )
 
@@ -17,7 +18,7 @@ type Store interface {
 	UpdateEntity(ctx context.Context, opt types.UpdateEntityOpt) error
 
 	// feature
-	CreateFeature(ctx context.Context, opt types.DBCreateFeatureOpt) error
+	CreateFeature(ctx context.Context, opt dbtypes.CreateFeatureOpt) error
 	GetFeature(ctx context.Context, featureName string) (*types.Feature, error)
 	ListFeature(ctx context.Context, groupName *string) ([]*types.Feature, error)
 	UpdateFeature(ctx context.Context, opt types.UpdateFeatureOpt) error
@@ -39,7 +40,7 @@ type Store interface {
 	GetRevision(ctx context.Context, groupName string, revision int64) (*types.Revision, error)
 	GetRevisionsByDataTables(ctx context.Context, dataTables []string) ([]*types.Revision, error)
 	BuildRevisionRanges(ctx context.Context, groupName string) ([]*types.RevisionRange, error)
-	InsertRevision(ctx context.Context, opt types.InsertRevisionOpt) error
+	InsertRevision(ctx context.Context, opt dbtypes.InsertRevisionOpt) error
 
 	io.Closer
 }
