+++ internal/database/metadata/postgres/feature.go
@@ -10,6 +10,7 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
 	"github.com/onestore-ai/onestore/internal/database"
+	dbtypes "github.com/onestore-ai/onestore/internal/database/types"
 	"github.com/onestore-ai/onestore/pkg/onestore/types"
 )
 
@@ -29,7 +30,7 @@ func (db *DB) validateDataType(ctx context.Context, dataType string) error {
 	})
 }
 
-func (db *DB) CreateFeature(ctx context.Context, opt types.DBCreateFeatureOpt) error {
+func (db *DB) CreateFeature(ctx context.Context, opt dbtypes.CreateFeatureOpt) error {
 	if err := db.validateDataType(ctx, opt.DBValueType); err != nil {
 		return fmt.Errorf("err when validating value_type input, details: %s", err.Error())
 	}
