+++ internal/database/offline/postgres/stream.go
@@ -5,10 +5,11 @@ import (
 	"fmt"
 	"strings"
 
+	dbtypes "github.com/onestore-ai/onestore/internal/database/types"
 	"github.com/onestore-ai/onestore/pkg/onestore/types"
 )
 
-func (db *DB) GetFeatureValuesStream(ctx context.Context, opt types.GetFeatureValuesStreamOpt) (<-chan *types.RawFeatureValueRecord, error) {
+func (db *DB) GetFeatureValuesStream(ctx context.Context, opt dbtypes.GetFeatureValuesStreamOpt) (<-chan *types.RawFeatureValueRecord, error) {
 	fields := append([]string{opt.EntityName}, opt.FeatureNames...)
 	query := fmt.Sprintf("select %s from %s", strings.Join(fields, ","), opt.DataTable)
 	if opt.Limit != nil {
